split out state functions
