+++ po/POTFILES.in
@@ -122,6 +122,7 @@ sidebar.c
 signal.c
 smtp.c
 sort.c
+state.c
 status.c
 system.c
 thread.c
